// src/types.ts
export interface Track {
  id: string;
  title: string;
  artist: string;
  url: string;
  duration: string;
  cover: string;
}

export type Point = { x: number; y: number };
export type Direction = 'UP' | 'DOWN' | 'LEFT' | 'RIGHT';

export interface GameState {
  snake: Point[];
  food: Point;
  direction: Direction;
  isGameOver: boolean;
  score: number;
  highScore: number;
}